Refactoring, moved some classes around
