+++ bn-apps/memberships-management/membership-service/src/main/kotlin/net/corda/businessnetworks/membership/Exceptions.kt
@@ -1,4 +1,4 @@
-package net.corda.businessnetworks.membership.common
+package net.corda.businessnetworks.membership
 
 import net.corda.businessnetworks.membership.states.Membership
 import net.corda.core.flows.FlowException
@@ -8,4 +8,3 @@ class NotAMemberException(val counterParty : Party) : FlowException("Counterpart
 class MembershipNotActiveException(val counterParty : Party) : FlowException("Counterparty's $counterParty membership in this business network is not active")
 class NotBNOException(val membership : Membership.State) : FlowException("This node is not the business network operator of this membership")
 class MembershipNotFound(val party : Party) : FlowException("Membership for party $party not found")
-class MultipleMembershipsFound(val party : Party) : FlowException("Multiple memberships found for party $party")
