+++ bn-apps/memberships-management/membership-service/src/main/kotlin/net/corda/businessnetworks/membership/bno/extension/MembershipAutoAcceptor.kt
@@ -3,7 +3,5 @@ package net.corda.businessnetworks.membership.bno.extension
 import net.corda.businessnetworks.membership.states.Membership
 
 interface MembershipAutoAcceptor {
-
     fun autoActivateThisMembership(membershipState : Membership.State) : Boolean
-
 }
