+++ bn-apps/memberships-management/membership-service/src/main/kotlin/net/corda/businessnetworks/membership/bno/support/BusinessNetworkOperatorFlowLogic.kt
@@ -1,35 +1,27 @@
 package net.corda.businessnetworks.membership.bno.support
 
 import net.corda.businessnetworks.membership.bno.service.DatabaseService
-import net.corda.businessnetworks.membership.common.MembershipNotFound
-import net.corda.businessnetworks.membership.common.MultipleMembershipsFound
-import net.corda.businessnetworks.membership.common.NotBNOException
+import net.corda.businessnetworks.membership.MembershipNotFound
+import net.corda.businessnetworks.membership.NotBNOException
 import net.corda.businessnetworks.membership.states.Membership
 import net.corda.core.contracts.StateAndRef
 import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.Party
-import net.corda.core.node.services.queryBy
 
 /**
  * Extend from this class if you are a business network operator and you want to make your life easier when writing
  * flows by getting access to the useful methods in this class.
  */
-abstract class BusinessNetworkOperatorFlowLogic<out T>() : FlowLogic<T>() {
+abstract class BusinessNetworkOperatorFlowLogic<out T> : FlowLogic<T>() {
 
     protected fun verifyThatWeAreBNO(membership : Membership.State) {
         if(ourIdentity != membership.bno) {
             throw NotBNOException(membership)
         }
     }
-
     protected fun findMembershipStateForParty(party : Party) : StateAndRef<Membership.State> {
-        //@todo this could be made more effective and look for the Party's state in the vault
-        val memberships = serviceHub.vaultService.queryBy<Membership.State>().states.filter { it.state.data.member == party }
-        return when {
-            memberships.isEmpty() -> throw MembershipNotFound(party)
-            memberships.size == 1 -> memberships.first()
-            else -> throw MultipleMembershipsFound(party)
-        }
+        val databaseService = serviceHub.cordaService(DatabaseService::class.java)
+        return databaseService.getMembership(party) ?: throw MembershipNotFound(party)
     }
 
     protected fun getActiveMembershipStates() : List<StateAndRef<Membership.State>> {
