--- bn-apps/memberships-management/membership-service/src/main/kotlin/net/corda/businessnetworks/membership/common/PartyAndMembershipMetadata.kt
@@ -1,8 +1,0 @@
-package net.corda.businessnetworks.membership.common
-
-import net.corda.businessnetworks.membership.states.MembershipMetadata
-import net.corda.core.identity.Party
-import net.corda.core.serialization.CordaSerializable
-
-@CordaSerializable
-data class PartyAndMembershipMetadata(val party : Party, val membershipMetadata: MembershipMetadata) {}
